Add "smarter" copying of code block snippets
Filter out the prompts from the code blocks so that the user
does not have to when copying using the copy buttons.

Remove the `lfe> ` and `$ ` that are prompts by seeing if they
are at the very beginning of the line.

Removed the copy buttons from the pure text blocks that
represent the results of shell commands or LFE expressions.

Also removed lines that start with LFE and Bash comments from
the snippet text if it is decided that the output should be
shown in the same block for scenarios like the guessing game,
where it might be convienent for multiple expressions to be copied
as a single chunk of text.